Greatly improve performance of Windows -> WSL clipboard synchronization
Previously win-to-clipboard functioned by embedding a Powershell script
which would track the active window. Through bash we would track if the
active window went from not-WSL to WSL. In this case, we would always
copy the clipboard contents. Copying the clipboard contents involved
starting a new powershell process which was very slow to copy.

Now, we do the not-WSL to WSL tracking from within the Powershell
script. To get more specific, when going from WSL to not-WSL, we keep
note of the current clipboard incremental value. Then, when going from
not-WSL to WSL, we see if the clipboard was incremented. If it was, we
output the clipboard contents as base64. On the WSL end, we consume the
base64 clipboard and decode it to populate the wayland clipboard. The
result is nearly instant clipboard synchronization.